cache: update to Go tip as of April 2023
As of commit 0fd6ae548f550bdbee4a434285ff052fb9dc7417.

Besides rewriting import paths, we swapped base.Fatalf with log.Fatalf,
and replaced cfg.Getenv with os.Getenv, adding a note about the
difference in behavior. The old code already had this limitation.

We hadn't updated this package since it was first copied in 2018,
so quite a few changes have taken place.
Of note, it now supports mmap; leave that out for now, to keep this
commit simple and to leave adding the mmap package for another patch.

A minor API change is that Trim now returns an error.
While technically a breaking change, the vast majority of users will be
simply calling the API without expecting a result, and that will
continue to work like it did before.
Checking for errors on trim is useful, which is why upstream added it.

Finally, the cache now uses lockedfile, which we already copied over.